@@ -1,6 +1,8 @@
 using System;
 using System.Net;
+using System.Net.Http;
 using System.Text.Json;
+using System.Threading.Tasks;
 
 namespace JishoNET.Models
 {
@@ -12,27 +14,62 @@ namespace JishoNET.Models
 		private static Uri BaseUrl = new Uri("https://jisho.org/api/v1/search/words?keyword=");
 
 		/// <summary>
-		/// Retrieve a a list of results from the Jisho API using the given keyword as a search term
+		/// Retrieve a list of results from the Jisho API using the given keyword as a search term asynchronously
 		/// </summary>
 		/// <param name="keyword">Keyword used as a search term to find definitions</param>
 		/// <returns><see cref="JishoResult" /> containing all definitions</returns>
-		public JishoResult GetDefinition(String keyword)
+		public async Task<JishoResult<JishoDefinition[]>> GetDefinitionAsync(String keyword)
 		{
 			try
 			{
-				WebClient webClient = new WebClient();
-				webClient.Headers.Add(HttpRequestHeader.UserAgent, "Mozilla/5.0");
-				webClient.Headers.Add(HttpRequestHeader.AcceptCharset, "UTF-8");
-				webClient.Encoding = System.Text.Encoding.UTF8;
-
-				String jsonResponse = webClient.DownloadString(new Uri(BaseUrl + keyword));
-				JishoResult result = JsonSerializer.Deserialize<JishoResult>(jsonResponse);
+				HttpClient client = new HttpClient();
+				var response = await client.GetAsync(BaseUrl + keyword);
+				JishoResult<JishoDefinition[]> result = JsonSerializer.Deserialize<JishoResult<JishoDefinition[]>>(response.Content.ReadAsStringAsync().Result);
+				result.Meta.Status = ((int)response.StatusCode);
 				result.Success = true;
 				return result;
 			}
 			catch (Exception e)
 			{
-				return new JishoResult()
+				return new JishoResult<JishoDefinition[]>()
+				{
+					Success = false,
+					Exception = e.ToString()
+				};
+			}
+		}
+
+		/// <summary>
+		/// Retrieve a list of results from the Jisho API using the given keyword as a search term
+		/// </summary>
+		/// <param name="keyword"></param>
+		/// <returns></returns>
+		public JishoResult<JishoDefinition[]> GetDefinition(String keyword)
+		{
+			return this.GetDefinitionAsync(keyword).Result;
+		}
+
+		/// <summary>
+		/// Quickly retrieve the top most result from Jisho using the given keyword as a search term asynchronously
+		/// </summary>
+		/// <param name="keyword">Keyword used as a search term to quickly retrieve an English <see cref="JishoEnglishSense" /> and a <see cref="JishoJapaneseDefinition" /> Reading</param>
+		/// <returns><see cref="JishoQuickDefinition" /> containing the top English <see cref="JishoEnglishSense" />  and <see cref="JishoJapaneseDefinition" /> Reading of the search term OR null if no definition was found</returns>
+		public async Task<JishoResult<JishoQuickDefinition>> GetQuickDefinitionAsync(String keyword)
+		{
+			try
+			{
+				JishoResult<JishoQuickDefinition> result = new JishoResult<JishoQuickDefinition>();
+				result = new JishoResult<JishoQuickDefinition>()
+				{
+					Data = new JishoQuickDefinition(await GetDefinitionAsync(keyword)),
+					Success = true,
+					Exception = null
+				};
+				return result;
+			}
+			catch (Exception e)
+			{
+				return new JishoResult<JishoQuickDefinition>()
 				{
 					Success = false,
 					Exception = e.ToString()
@@ -43,25 +80,11 @@ namespace JishoNET.Models
 		/// <summary>
 		/// Quickly retrieve the top most result from Jisho using the given keyword as a search term
 		/// </summary>
-		/// <param name="keyword">Keyword used as a search term to quickly retrieve an English <see cref="JishoEnglishSense" /> and a <see cref="JishoJapaneseDefinition" /> Reading</param>
-		/// <returns><see cref="JishoQuickDefinition" /> containing the top English <see cref="JishoEnglishSense" />  and <see cref="JishoJapaneseDefinition" /> Reading of the search term OR null if no definition was found</returns>
-		public JishoQuickDefinition GetQuickDefinition(String keyword)
+		/// <param name="keyword"></param>
+		/// <returns></returns>
+		public JishoResult<JishoQuickDefinition> GetQuickDefinition(String keyword)
 		{
-			try
-			{
-				JishoQuickDefinition result = new JishoQuickDefinition();
-				result = new JishoQuickDefinition(GetDefinition(keyword));
-				result.Success = true;
-				return result;
-			}
-			catch (Exception e)
-			{
-				return new JishoQuickDefinition()
-				{
-					Success = false,
-					Exception = e.ToString()
-				};
-			}
+			return this.GetQuickDefinitionAsync(keyword).Result;
 		}
 	}
 }
